Move and update `TimeFieldtype` rule class to `Statamic\Rules` w/ Laravel conventions.
