+++ src/Rules/TimeFieldtype.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace Statamic\Validation;
+namespace Statamic\Rules;
 
+use Closure;
 use DateTime;
-use Illuminate\Contracts\Validation\InvokableRule;
+use Illuminate\Contracts\Validation\ValidationRule;
 
-class TimeFieldtype implements InvokableRule
+class TimeFieldtype implements ValidationRule
 {
     protected $fieldtype;
 
@@ -14,12 +15,12 @@ class TimeFieldtype implements InvokableRule
         $this->fieldtype = $fieldtype;
     }
 
-    public function __invoke($attribute, $value, $fail)
+    public function validate(string $attribute, mixed $value, Closure $fail): void
     {
         $format = $this->fieldtype->config('seconds_enabled') ? 'H:i:s' : 'H:i';
 
         if (! $this->matchesFormat($value, $format)) {
-            return $fail('statamic::validation.time')->translate();
+            $fail('statamic::validation.time')->translate();
         }
     }
