+++ tests/Fieldtypes/TimeTest.php
@@ -62,18 +62,22 @@ class TimeTest extends TestCase
      *
      * @dataProvider validationProvider
      */
-    public function it_validates($config, $input, $expected)
+    public function it_validates($config, $input, $passes)
     {
         $field = $this->fieldtype($config)->field();
-        $messages = [];
+        $messages = collect();
 
         try {
             Validator::validate(['test' => $input], $field->rules(), [], $field->validationAttributes());
         } catch (ValidationException $e) {
-            $messages = $e->validator->errors()->all();
+            $messages = $e->validator->errors();
         }
 
-        $this->assertEquals($expected, $messages);
+        if ($passes) {
+            $this->assertCount(0, $messages);
+        } else {
+            $this->assertEquals(__('statamic::validation.time'), $messages->first());
+        }
     }
 
     public static function validationProvider()
@@ -82,37 +86,37 @@ class TimeTest extends TestCase
             'valid time' => [
                 [],
                 '14:00',
-                [],
+                true,
             ],
             'valid time with seconds' => [
                 ['seconds_enabled' => true],
                 '14:00:00',
-                [],
+                true,
             ],
             'invalid time format' => [
                 [],
                 'not formatted like a time',
-                ['Not a valid time.'],
+                false,
             ],
             '12 hour time' => [
                 [],
                 '1:00',
-                ['Not a valid time.'],
+                false,
             ],
             'invalid hour' => [
                 [],
                 '25:00',
-                ['Not a valid time.'],
+                false,
             ],
             'invalid minute' => [
                 [],
                 '14:65',
-                ['Not a valid time.'],
+                false,
             ],
             'invalid second' => [
                 ['seconds_enabled' => true],
                 '13:00:60',
-                ['Not a valid time.'],
+                false,
             ],
         ];
     }
